requirements.txt
matplotlib~=3.9.0
numpy~=1.26.4
scipy~=1.13.1
pillow~=10.3.0
opencv-python~=4.9.0.80